Shift repeat logic to its own file
